OldFiles_esg-ai-pipeline/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.3
openpyxl==3.1.2
azure-ai-documentintelligence==1.0.0
openai==1.3.0
pydantic==2.5.0
aiofiles==23.2.1
Pillow==10.1.0
